requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
six==1.15.0
websockets==8.1
yarl==1.4.2
youtube-dl==2020.5.29
